ByteSource overloads for all media sends; PART/BIG_FILE as Long.

+++ core/src/commonMain/kotlin/iris/kmtproto/api/user/Upload.kt
@@ -17,8 +17,8 @@ import kotlinx.coroutines.Deferred
 
 /** `upload.saveFilePart` / `upload.saveBigFilePart`. */
 object Upload {
-    internal const val PART = 512 * 1024
-    internal const val BIG_FILE = 10 * 1024 * 1024
+    internal const val PART = 512 * 1024L
+    internal const val BIG_FILE = 10 * 1024 * 1024L
 
     fun saveFileAsync(client: TelegramClient, bytes: ByteArray, name: String): Deferred<InputFile> =
         client.apiAsync { saveFile(client, bytes, name) }
@@ -36,11 +36,11 @@ object Upload {
         val parts = ((size + PART - 1) / PART).toInt()
         val fileId = nextFileId()
         val hasher = if (big) null else Md5Hasher()
-        val buf = ByteArray(PART)
+        val buf = ByteArray(PART.toInt())
         var remaining = size
         var index = 0
         while (remaining > 0L) {
-            val want = minOf(PART.toLong(), remaining).toInt()
+            val want = minOf(PART, remaining).toInt()
             var filled = 0
             while (filled < want) {
                 val n = source.read(buf, filled, want - filled)
@@ -48,7 +48,7 @@ object Upload {
                 filled += n
             }
             hasher?.update(buf, 0, filled)
-            val chunk = if (filled == PART) buf else buf.copyOf(filled)
+            val chunk = if (filled.toLong() == PART) buf else buf.copyOf(filled)
             val ok = if (big) {
                 client.invoke(
                     UploadSaveBigFilePart(
